Avoid fetching bounds multiple times in the MID_POINT ValueEstimator
When using histograms, we see fairly high allocations and survivors
when reading quantiles with the MID_POINT ValueEstimator.

We should be able to avoid some if we do not retrieve the upper and
lower bound twice for each access.

Allocations actually happen in AbstractLayout$$Lambda$.get$Lambda
because the lambda-function there needs to access the outer-scope
and thus cannot be optimized away. Avoiding the lambda-function
would potentially improve allocations a lot, but it is not done
as part of this commit.

(cherry picked from commit 47ac638e1cff06191c53f142fb7c918f6e78c320)